requirements.txt
# 核心框架
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.12

# 数据源和API
requests>=2.31.0
newsapi-python>=0.2.6
praw>=7.7.1
feedparser>=6.0.10

# 网页解析
beautifulsoup4>=4.12.2
lxml>=4.9.3

# 内容处理
openai>=1.3.0
transformers>=4.35.0
numpy>=1.24.0,<2.0.0  # 限制版本避免兼容性问题
pandas>=2.0.0

# 文本预处理和NLP
jieba>=0.42.1
nltk>=3.8.1
scikit-learn>=1.3.0
regex>=2023.10.3
langdetect>=1.0.9
difflib2>=0.1.0

# 用户界面
gradio>=4.0.0

# 邮件服务
sendgrid>=6.10.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 实用工具
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0

# 测试和开发
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# 日志
loguru>=0.7.2 